third version

+++ leave/src/main/java/com/lms/dto/response/LeaveTypeBreakdown.java
@@ -1,17 +1,13 @@
 package com.lms.dto.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class LeaveTypeBreakdown {
-
     private String leaveType;
-    private double allocatedDays;
-    private double usedDays;
-    private double remainingDays;
-    private long halfDayCount;
+    private Double allocatedDays;
+    private Double usedDays;
+    private Double remainingDays;
+    private Integer halfDayCount;
 }
